requirements.txt
requests~=2.32.3
pybit~=5.7.0
pandas~=2.2.2
ta~=0.11.0
python-dotenv~=1.0.1
pycryptodome==3.10.1